:fire: Add Pipelines that allow user to manipulate changes.  :wrench: Fix typo in method name on ActivityLog options.  :twisted_rightwards_arrows: Merge DetectsChanges trait into LogsActivity.  :recycle: Clean some unnecessary helper functions.  :memo: Document some behaviours.  :white_check_mark: Add 2 new tests to demonstrate powerfulness of the pipeline.
