+++ src/EventLogBag.php
@@ -1,0 +1,18 @@
+<?php
+namespace Spatie\Activitylog;
+
+use Illuminate\Database\Eloquent\Model;
+
+class EventLogBag
+{
+    public function __construct(
+        public string $name,
+        public Model $model,
+        public array $changes,
+        public ?ActivitylogOptions $options = null
+    ) {
+        if (is_null($this->options)) {
+            $this->options = $model->getActivitylogOptions();
+        }
+    }
+}
