+++ tests/Models/Issue733.php
@@ -16,7 +16,7 @@ class Issue733 extends Article
     public function getActivitylogOptions() : ActivitylogOptions
     {
         return ActivitylogOptions::create()
-        ->DontSubmitEmptyLogs()
+        ->dontSubmitEmptyLogs()
         ->logOnly(['name']);
     }
 }
